refactor: start refactoring markdown components
