+++ package.json
@@ -46,8 +46,6 @@
     "@radix-ui/react-slot": "^1.0.2",
     "@sentry/react": "^7.43.0",
     "@sentry/tracing": "^7.43.0",
-    "@uiw/react-markdown-preview": "^4.1.15",
-    "@uiw/react-md-editor": "^3.14.3",
     "@wagmi/core": "^1.3.9",
     "bignumber.js": "^9.1.1",
     "buffer": "^6.0.3",
