--- src/components/Markdown/Editor/styles-overrides.css
@@ -1,47 +1,0 @@
-.wmde-markdown,
-.wmde-markdown-var {
-  color-scheme: dark;
-  --color-prettylights-syntax-comment: #8b949e;
-  --color-prettylights-syntax-constant: #79c0ff;
-  --color-prettylights-syntax-entity: #d2a8ff;
-  --color-prettylights-syntax-storage-modifier-import: #c9d1d9;
-  --color-prettylights-syntax-entity-tag: #7ee787;
-  --color-prettylights-syntax-keyword: #ff7b72;
-  --color-prettylights-syntax-string: #a5d6ff;
-  --color-prettylights-syntax-variable: #ffa657;
-  --color-prettylights-syntax-brackethighlighter-unmatched: #f85149;
-  --color-prettylights-syntax-invalid-illegal-text: #f0f6fc;
-  --color-prettylights-syntax-invalid-illegal-bg: #8e1519;
-  --color-prettylights-syntax-carriage-return-text: #f0f6fc;
-  --color-prettylights-syntax-carriage-return-bg: #b62324;
-  --color-prettylights-syntax-string-regexp: #7ee787;
-  --color-prettylights-syntax-markup-list: #f2cc60;
-  --color-prettylights-syntax-markup-heading: #1f6feb;
-  --color-prettylights-syntax-markup-italic: #c9d1d9;
-  --color-prettylights-syntax-markup-bold: #c9d1d9;
-  --color-prettylights-syntax-markup-deleted-text: #ffdcd7;
-  --color-prettylights-syntax-markup-deleted-bg: #67060c;
-  --color-prettylights-syntax-markup-inserted-text: #aff5b4;
-  --color-prettylights-syntax-markup-inserted-bg: #033a16;
-  --color-prettylights-syntax-markup-changed-text: #ffdfb6;
-  --color-prettylights-syntax-markup-changed-bg: #5a1e02;
-  --color-prettylights-syntax-markup-ignored-text: #c9d1d9;
-  --color-prettylights-syntax-markup-ignored-bg: #1158c7;
-  --color-prettylights-syntax-meta-diff-range: #d2a8ff;
-  --color-prettylights-syntax-brackethighlighter-angle: #8b949e;
-  --color-prettylights-syntax-sublimelinter-gutter-mark: #484f58;
-  --color-prettylights-syntax-constant-other-reference-link: #a5d6ff;
-  // Custom
-  --color-fg-default: rgba(255, 255, 255, 1);
-  --color-fg-muted: rgba(161, 161, 161, 1);
-  --color-fg-subtle: rgba(149, 151, 161, 1);
-  --color-canvas-default: rgba(255, 255, 255, 1);
-  --color-canvas-subtle: rgba(22, 27, 39, 1);
-  --color-border-default: rgba(49, 50, 60, 1);
-  --color-border-muted: rgba(49, 50, 60, 1);
-  --color-neutral-muted: rgba(32, 39, 56, 1); // hover background
-  --color-accent-fg: rgba(58, 120, 255, 1); // hover
-  --color-accent-emphasis: rgba(58, 120, 255, 1);
-  --color-attention-subtle: rgba(245, 120, 66, 1);
-  --color-danger-fg: rgba(233, 61, 102, 1);
-}
