--- src/components/Markdown/index.ts
@@ -1,3 +1,0 @@
-export { default as MarkdownEditor } from './Editor';
-export * from './Editor';
-export { default as MarkdownViewer } from './Viewer';
