+++ src/components/MarkdownViewer/index.stories.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 
 import { withCenterStory } from 'stories/decorators';
 
-import MarkdownViewer from '.';
+import { MarkdownViewer } from '.';
 
 export default {
   title: 'Components/Markdown/Viewer',
