+++ src/components/index.ts
@@ -37,7 +37,6 @@ export * from './TokenIconWithSymbol';
 export * from './TokenTextField';
 export * from './Tooltip';
 export * from './ValueUpdate';
-export * from './Markdown';
 export * from './SelectTokenTextField';
 export * from './InfoIcon';
 export * from './CellGroup';
@@ -50,3 +49,5 @@ export * from './SpendingLimit';
 export * from './Link';
 export * from './AuthModal';
 export * from './TokenAnnouncement';
+export * from './MarkdownEditor';
+export * from './MarkdownViewer';
